Go back to Buffer usage flags instead of type;
For compute skinning we can't easily have a Buffer used for storage and
vertex.  Usage flags are better for this.  I think alignment issues can
be solved using lcm(blockAlign, vertexAlign) which can waste lots of
space in some cases but is easier than putting a PO2 limit on people or
greatly restricting buffer usages.