Fix possible gltf crash;
blobCount/imageCount are set before their arrays are allocated, so we
need to see if the arrays exist before we try to clean them up.

It doesn't matter if all the array contents have been filled in yet,
since their values will be NULL and this is safe to pass to lovrRelease.